data_breach_backend/requirements/base.txt
Django==3.2.9
django-cors-headers==3.10.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.91
gunicorn==19.9.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
six==1.16.0
